requirements.txt
boto==2.49.0
boto3==1.9.247
botocore==1.12.247
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-redis==4.10.0
django-storages==1.7.2
docutils==0.15.2
gunicorn==19.9.0
jmespath==0.9.4
Pillow==6.2.0
pipfile==0.0.2
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.3
redis==3.3.10
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.6
whitenoise==4.1.4
